move feature_extraction tests as well
